requirements/base.txt
Django==1.5.1
argparse==1.2.1
boto==2.9.2
dj-database-url==0.2.1
django-storages==1.1.8
gunicorn==0.17.4
html5lib==0.95
isodate==0.4.9
psycopg2==2.5
rdflib==3.4.0
requests==1.2.0
trollop==0.0.11
wsgiref==0.1.2
django-suit==0.2.4
django-celery==3.0.21
redis==2.7.6